BackEnd/Nothing/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.13.0
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
Pillow==10.1.0
sqlparse==0.2.4
